Fixed gallery update issue after importing
